refactor tests to remove __init__ files useless imports
