--- tests/validator/__init__.py
@@ -1,4 +1,0 @@
-##################################################################
-# Copyright 2018 Open Source Geospatial Foundation and others    #
-# licensed under MIT, Please consult LICENSE.txt for details     #
-##################################################################
